+++ OpenSim/Common/Test/testComponentInterface.cpp
@@ -1634,16 +1634,21 @@ void testTableReporter() {
             TimeStepper ts(system, integ);
             ts.initialize(s);
             ts.stepTo(1.0);
-            std::cout << "TableReporter table after simulating:\n";
             const auto& table = reporter->getTable();
-            SimTK_TEST_MUST_THROW_EXC(table.toString(), EmptyTable);
+            std::cout << "TableReporter table after simulating:\n"
+                      << table.toString() << std::endl;
+            SimTK_TEST_MUST_THROW_EXC(table.getDependentColumnAtIndex(0),
+                                      EmptyTable);
         }
 
         // Ensure that clearing the table and performing a new simulation works
         // even if the reporter's input has no connectees.
         reporter->clearTable();
-        std::cout << "TableReporter table after clearing:\n";
-        SimTK_TEST_MUST_THROW_EXC(reporter->getTable().toString(), EmptyTable);
+        std::cout << "TableReporter table after clearing:\n"
+                  << reporter->getTable().toString() << std::endl;
+        SimTK_TEST_MUST_THROW_EXC(
+            reporter->getTable().getDependentColumnAtIndex(0),
+            EmptyTable);
     
         {
             SimTK::State s = system.realizeTopology();
@@ -1652,9 +1657,11 @@ void testTableReporter() {
             TimeStepper ts(system, integ);
             ts.initialize(s);
             ts.stepTo(1.0);
-            std::cout << "TableReporter table after simulating again:\n";
             const auto& table = reporter->getTable();
-            SimTK_TEST_MUST_THROW_EXC(table.toString(), EmptyTable);
+            std::cout << "TableReporter table after simulating again:\n"
+                      << table.toString() << std::endl;
+            SimTK_TEST_MUST_THROW_EXC(table.getDependentColumnAtIndex(0),
+                                      EmptyTable);
         }
     }
 }