+++ OpenSim/Simulation/Model/TwoFrameLinker.h
@@ -607,7 +607,8 @@ void TwoFrameLinker<C, F>::updateFromXMLNode(SimTK::Xml::Element& aNode,
     int documentVersion = versionNumber;
     if (documentVersion < XMLDocument::getLatestVersion()) {
         if (documentVersion < 30505) {
-            // replace old properties with latest use of PhysicalOffsetFrames properties
+            // replace old properties with latest use of PhysicalOffsetFrames
+            // properties
             SimTK::Xml::element_iterator body1Element =
                 aNode.element_begin("body_1");
             SimTK::Xml::element_iterator body2Element =
@@ -633,15 +634,11 @@ void TwoFrameLinker<C, F>::updateFromXMLNode(SimTK::Xml::Element& aNode,
                 body2Element->getValueAs<std::string>(frame2Name);
             }
 
-            XMLDocument::addConnector(aNode, "Connector_PhysicalFrame_",
-                "frame1", frame1Name);
-            XMLDocument::addConnector(aNode, "Connector_PhysicalFrame_",
-                "frame2", frame2Name);
-
             Vec3 locationInFrame1(0);
             Vec3 orientationInFrame1(0);
             Vec3 locationInFrame2(0);
             Vec3 orientationInFrame2(0);
+
 
             if (locBody1Elt != aNode.element_end()) {
                 locBody1Elt->getValueAs<Vec3>(locationInFrame1);
@@ -656,22 +653,46 @@ void TwoFrameLinker<C, F>::updateFromXMLNode(SimTK::Xml::Element& aNode,
                 orientBody2Elt->getValueAs<Vec3>(orientationInFrame2);
             }
 
+            // The value of the connectee name depends on if we need to insert
+            // offset frames.
+            std::string frame1_connectee_name;
+            std::string frame2_connectee_name;
+
             // now append updated frames to the property list if they are not
             // identity transforms.
             if ((locationInFrame1.norm() > 0.0) ||
                 (orientationInFrame1.norm() > 0.0)) {
-                XMLDocument::addPhysicalOffsetFrame(aNode, frame1Name + "_offset",
-                    frame1Name, locationInFrame1, orientationInFrame1);
-                body1Element->setValue(frame1Name + "_offset");
+                frame1_connectee_name = frame1Name + "_offset";
+                XMLDocument::addPhysicalOffsetFrame30505(aNode,
+                        frame1_connectee_name,
+                        frame1Name, locationInFrame1, orientationInFrame1);
+            } else {
+                // In pre-4.0 models, BushingForces, Constraints, and all other
+                // classes upgraded to use TwoFrameLinker, and the
+                // Bodies they depended on, are necessarily 1 level deep
+                // (subcomponent of model). Therefore, prepend "../" to get the
+                // correct relative path.
+                frame1_connectee_name = "../" + frame1Name;
             }
 
             // again for the offset frame on the child
             if ((locationInFrame2.norm() > 0.0) ||
                 (orientationInFrame2.norm() > 0.0)) {
-                XMLDocument::addPhysicalOffsetFrame(aNode, frame2Name + "_offset",
-                    frame2Name, locationInFrame2, orientationInFrame2);
+                frame2_connectee_name = frame2Name + "_offset";
+                XMLDocument::addPhysicalOffsetFrame30505(aNode,
+                        frame2_connectee_name,
+                        frame2Name, locationInFrame2, orientationInFrame2);
                 body2Element->setValue(frame2Name + "_offset");
+            } else {
+                frame2_connectee_name = "../" + frame2Name;
             }
+
+            // Now we know whether to use the original body_1 and body_2
+            // strings or if we should use the offsets.
+            XMLDocument::addConnector(aNode, "Connector_PhysicalFrame_",
+                "frame1", frame1_connectee_name);
+            XMLDocument::addConnector(aNode, "Connector_PhysicalFrame_",
+                "frame2", frame2_connectee_name);
         }
     }
     Super::updateFromXMLNode(aNode, versionNumber);